Move index.js to root of package
Changes:
* See topic line

Why?
* It makes importing and using this package more intuitive :D
